Added VelocityGraph tests. Fixed a stack overflow bug in ElementHelper.AreEqual. Got rid of float conversions.

+++ Blueprints/blueprints-core/Util/IO/GML/GMLParser.cs
@@ -178,7 +178,7 @@ namespace Frontenac.Blueprints.Util.IO.GML
                         
                         if (doubleValue == Math.Floor(doubleValue))
                             return (int)doubleValue;
-                        return (float)doubleValue;
+                        return doubleValue;
                     }
                     if (type == '[')
                         return ParseMap(key, st);
